feat(filetypes): Prompt users to install YAML extension for Buildspec and Devfiles
Problem:
- Currently our yaml activation doesn't check for buildspec or devfiles so users may not be exposed to those features if they don't already have VSCode-YAML installed

Solution:
- Prompt for YAML extension activation on buildspec and devfiles. The only differences in this implementation is that we now watch on multiple different filenames and we also use the cfn/buildspec registry to determine if the prompt should activate
