+++ .changes/next-release/Feature-8cc2b8d7-860e-455e-aa73-01481d00f74c.json
@@ -1,0 +1,4 @@
+{
+	"type": "Feature",
+	"description": "Prompt users to install YAML extension for Buildspec and Devfiles"
+}
